% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_summary}
\alias{plot_summary}
\title{Visualize the summary of a data frame}
\usage{
plot_summary(x, base = c("raw", "hist"), if_box = FALSE, if_message = FALSE)
}
\arguments{
\item{x}{Data frame}

\item{base}{Character. "raw" or "hist"}

\item{if_box}{Logical. Whether plot the box}

\item{if_message}{Logical. Whether display the progress message}
}
\value{
a figure
}
\description{
Visualize the summary of a data frame
}
\examples{
plot_summary(iris)
plot_summary(iris, if_box = TRUE)
}
